feat: move `databases` commands under `server`
