+++ src/argilla/__main__.py
@@ -16,7 +16,6 @@
 import warnings
 
 from argilla.tasks import (
-    database_app,
     datasets_app,
     info_app,
     login_app,
@@ -33,7 +32,6 @@ warnings.simplefilter("ignore", UserWarning)
 
 app = AsyncTyper(rich_help_panel=True, help="Argilla CLI", no_args_is_help=True)
 
-app.add_typer(database_app, name="database")
 app.add_typer(datasets_app, name="datasets")
 app.add_typer(info_app, name="info")
 app.add_typer(login_app, name="login")
